// src/app/dtos/global/agent.dto.ts
export interface AgentDto {
  montantTraitementsSalaires: string;
  montantCotisationsSociales: string;
  montantPrestationsFamiliales: string;
  montantPrisesChargeMedicales: string;
  montantAutresChargesPersonnel: string;
  montantPrimes: string;
  montantIndemnites: string;
  montantContractuels: string;
  chapitre: string;
  codeAction: string;
  libAction: string;
  codeActivite: string;
  libActivite: string;
  matricule: string;
  nom: string;
  emploi: string;
  statut: string;
  age: number;
}